[controller] [admin-tool] Max Record Size Config (#1027)
Added a new store-level config setting `maxRecordSizeBytes`. This will be used to decide if a batch push job should fail or if nearline consumption should be paused due the ingestion of records that are too large.
  * The default value of `maxRecordSizeBytes` is `-1`. By leaving the limit as unset by default, it is easier for us to add a global config in the future should we deem necessary.
  * This default value of `maxRecordSizeBytes` will be changed to `10MB` in-memory when it reaches its eventual destination at `VeniceWriter`, even if the config is set to `-1`.
  * The adjustment is at the store-level, as we anticipate only a few stores could surpass the threshold and require an increased limit.

Added tests `testUpdateMaxRecordSize()` and `testChunkingMaxRecordSize()` in `VeniceParentHelixAdminTest`. The former verifies that the config is read and set properly, and the latter
verifies that the default value of `maxRecordSizeBytes` will be `10MB` if chunking is enabled.

Refactored tests involving "setting a value and verifying the value was set" into a reusable function `testUpdateConfig()`. All you need to do is provide `Consumer` functions for setting a value and verifying the value, which will hopefully make testing configs very straightforward.