[Specs] Add TLA+ Specification for a Bounded Collection Operation
Today in Venice we support CRDT based operations on collections for write compute.  These operations include the ability to add to a collection, remove from a collection, and set a collection to a value.  An important property of this data type is that these operations are commutative, and can be replayed in any order and come to a consistent result.  This is important for AA where two datacenters may apply the same operations in any order.

One problem we currently suffer from is that there is no bound to the size of a collection field in Venice.  Since writers are working on an entry at a time, we have situations where this record can grow to a very large size, leading to expensive update and read operations in the database.

This spec introduces the notion of a bounded collection which has a max size and retains the top N elements.

For a bounded collection, it would be ideal for this collection to have the same operations available to an unbounded collection, with the addition of the max size.  This spec explores that idea, and makes an interesting observation.

Adding and removing elements with a tombstone list of an unbounded collection is commutative.

Adding and removing elements with a tombstone list and a max size is NOT commutative.

The spec as published in this PR fails model checking.  It can pass model checking with some tweaks to the cfg file (setting MAX_SIZE to less then zero removes the bound and makes it pass, also having no DELETE operations makes the checker pass).