[server] Fix a race condition in SIT graceful shutdown workflow
In defered write mode, a running checksum is calculated by drainer for every
key/value pair from Kafka PUT message and then stored in PartitionConsumptionState.expectedSSTFileChecksum.

When we gracefully close a StoreIngestionTask, SIT thread calls syncOffset to checkpoint
current state and the underlying RocksDBSstFileWriter who executes the sync call also
calculates checksum of all the key/value pair stored in the input sstFilePath and then
verifies if the checksum matches with the value stored in PartitionConsumptionState.

Because the above actions happen in different threads, a race can happen when they run
concurrently and a VeniceChecksumException will be thrown in SIT.

To fix this issue, this PR doesn't allow SIT thread to call syncOffset directly.
Instead, we let SIT thread to insert a SYNC_OFFSET command node into the drainer
queue and let the same drainer thread to execute the SYNC_OFFSET action so that
we don't have any race.

SIT thread then waits for command to finish within a certiain time period.
If times out, SIT can invalidate the command if it hasn't been executed yet.
Otherwise, if the command is alread in execution, SIT continues to wait for
it to finish.