[vpj][server] Fix DELETE handling in TTL repush and allow server to auto convert value level TS to field level TS (#1081)
This PR addresses two issues related to AAWC:
(1) SIT does not fail when it tries to merge PUT with existing value level RMD record. This can happen when user enables PartialUpdate for an AA store and performs either a restart or repush.
(2) TTL Repush job does not fail when DELETE is the latest value of a key.