[server] Fixed ServerStoreAclHandler unable to process DICTIONARY query action
In venice server the resource for ServerStoreAclHandler is a version topic for STORAGE and COMPUTE query
actions but it's still a store name if the query action is DICTIONARY. Always processing it as a version
topic will result in an empty string as the resource name.

Added tests with a mock DynamicAccessController to verify basic request parsing and handling for
ServerStoreAclHandler.