[server] Added a throttler for AAWC workload of leader replica
We know AA/WC workload is expensive and in the past, we were
trying to limit the consumer pool for AA/WC leader to control the
resource usage, but sometimes, it is not very accurate.
This code change introduces a new throttler for Leader AA/WC workload,
and it will only take effect when the dedicated consumer pool for
AA/WC leader feature is enabled.
New config:
server.aa.wc.leader.quota.records.per.second : default (-1)