Bump Jackson Version to 2.13.4 (#1109)
Bump Jackson Version to 2.13.4 to resolve ELR failure

Error message: In FasterXML jackson-databind before 2.12.7.1 and in 2.13.x before 2.13.4, resource exhaustion can occur because of a lack of a check in BeanDeserializer._deserializeFromArray to prevent use of deeply nested arrays. An application is vulnerable only with certain customized choices for deserialization.