[test] Microbenchmark for TokenBucket
We suspected that TokenBucket may be slow, though this microbenchmark indicates that
it can run at ~microsecond latency even under high concurrency, and at sub-microsec
under light load. So this is not the bottleneck we were looking for, but we'll keep
the microbench checked in just in case we need it again in the future.