[doc] Venice Large Record Value Chunking Diagram (#1107)
Write-path diagram for large record value chunking in Venice. There is a link to the source diagram on Whimsical, and you'll need to copy the diagram to your account in order to edit it. Also included an svg file copy from Whimsical's experimental feature, but it's kinda glitchy and missing some lines.