[server] print out slow consumer thread backtrace (#1120)
* [server] print out slow consumer thread backtrace

Today, when a slow consumer thread is found, it normally requires extra manual steps for debugging,
e.g. requesting a thread dump, searching the slowed thread from the dump etc. Such manual steps could
slow down the debugging process itself and even worse, at the moment when a thread was taken, the slow
thread may already passed the jamming point, so it is hard to know exactly where it was slowed down.

This PR adds a stack trace print out message when Venice believes that a slow consumer thread is found.