[dvc] Fixed DVC integration test resource leak (#1147)
There is a resource leak in DaVinciClientTest::testBlobP2PTransferAmongDVC which
causes subsequent tests to fail. This is occasionally masked by the test-retry
plugin, though the retry itself is flaky and so it sometimes fails to retry at
all. In any case, it is better to not fail in the first place...

Also did some minor tweaks to the main code to make debugging issues easier in
the future. This includes extra logging in DaVinciBackend and a slightly more
precise exception message in AvroGenericDaVinciClient.