[router][samza] Harden up quota monitor requests (#1148)
Main fix is that there exists a condition in the router where during a store move, the store will have information in the routers internal store config repository, but it won't have it in the store repository after the store was deleted at the end of the store move.  This would then trigger an exception and 500 error, and would prevent producers from reinitializing their clients. This fix addresses that.

Added tests in the store migration test suite to include a running producer.  Also added some protection in the code against empty responses (which was encountered while running the debugger).