[server][router] Handle migrated store before checking ACL (#1175)
When a store is completely migrated, but there is still some client that didn't have queries at time of migration, and hence didn't migrate to the new cluster. We fixed this issue in #843, but removed the logic to check the store existence from the ACL handler with the expectation that the next layers would redirect the requests for migrated store.

The interface of the ACL handler doesn't specify how to handle the case of a missing store. So, it is possible that the ACL handler could either:
1. Return "false" to the "hasAccess" call
2. Throw an "AclException"

Only in the second case would the request be forwarded to the next handler. This commit adds the store existence check before checking ACLs and hence, can forward the request to the next handler so that the client can be notified of the store move.

This commit unifies common store ACL handler logic into "AbstractStoreAclHandler" and splits router-specific and server-specific logic into their independent implementations: "RouterStoreAclHandler" and "ServerStoreAclHandler"