[fast-client] Replace per-store per-route tracking by per-route tracking
This PR is trying to reduce the total number of metrics emitted by
fast-client.
Previously, Fast-Client will have the instance-level tracking
for each store, which will create a lot of metrics when the application
is using multiple stores, and this code change will only emit
tacking metrics (qps, latency) in instance level rather than
per store per instance.

TODO:
There still is a metric, which is tracking pending request count
per store per route and it requires additional effort to change
this mechanism as today's `InstanceHealthMonitor` is per store
per instance and in theory, it should be per instance, and we
will use another PR to address it.