[controller] Stop writing writing ready-to-serve instances info in meta system store (#1228)
When materializing the meta store, we generate a snapshot of ready to serve instances  
for all user store versions. However, if a version is in an ERROR state, its Helix resources  
may be absent. Checking these resources in such cases can trigger exceptions,  
potentially interrupting the addVersion process.

Previously, the meta store required readyToServe instances data for client use,
but since no clients currently rely on it, it's safe to discontinue storing this
information in the meta system store.