[controller] Refresh OfflinePushStatus after change listener subscription during loadAllPushes() call (#1217)
The issue:
We observed such race condition in controller deployment/restart. When it restarts, it will call loadAllPushes() which load all OfflinePushStatus snapshot for all topics from ZK, it will then iterate the list and try to subscribe to the changes. If there is any change in between, it will be missed and push job can hang forever.

This PR adds a refresh offline pushes call after listener subscription. It is using the same store level lock, so it should not cause any race condition or deadlock with change listener callback.