[controller][server][dvc][vpj][producer][consumer] Fixed sys store init (#1238)
The SystemStoreInitializationHelper, when running in parent controller 
mode, would keep creating new versions of system stores, thus leading 
to data loss in those system stores. This commit changes to check to
make it idempotent.

Also bumped up the Kafka dependency to 2.4.1.78. This fixes in an issue 
in NetworkClient::leastLoadedNode which manifested as a flood of 
IllegalArgumentException when some Kafka partition became completely 
unavailable.

Miscellaneous test changes:

- Tactical changes to attempt to mitigate DaVinciClientMemoryLimitTest's
  flakiness. It unfortunately does not address the RC of the flakiness, 
  however.

- Also bumped up the RF to 2 on some flaky integration tests to see if it 
  helps.