[server] Fixed storage node read quota usage ratio spikes
1. We observed unusually high quota usage ratio spikes (7x) without any rejected requests. This is mathematically
impossible if the spikes were caused by an actual KPS spike since our default sampling window is 30s and token
bucket capacity multiplier is only 5x. This is only possible when the node responisbility (divisor) changed and
we are still using the "old" kps to calculate the usage ratio.

2. Another issue was that node responsibility could be updated whenever partition assignment changed, regardless
if the change was actually for the serving/current version. This is bad since it will not only fluctuate the usage
ratio in unexpected ways but also making it inaccurate.