[controller] Set min RF on controllers and storage clusters to RF - 1 with a floor of 1 (#1242)
Currently, the min replication factor (RF) for controllers is equal to replication factor. During a deployment, the number of instances could go below what is set to RF, which would cause a rebalance failure when using Waged.

To prevent this, we will be setting min RF = max(RF - 1, 1). This will be applied to the storage clusters too.