[controller] Add instance stoppable check based on CV states (#1198)
Implement customized server stoppable check API in controller

The customized stoppable check takes a list of instances, along with a list of previously approved instances and decide which of these instances can be stopped.

For the instances that cannot be stopped, the controller returns an enum that denotes the reason for rejecting the request. Also move existing pre-stop hook of `isRemovable` to use CV instead of EV for checking the replica status as CV is what determines what is being served in the read path.