[router] Added more debugging details to compression inconsistency issue (#1253)
The router expects to get the same compression scheme from all server
responses when performing a batch get, but it seems that this is not
always the case. This change will log the response headers in case this
happens, in the hope that it gives more clues about what went wrong.