[controller] Set Helix instance tags automatically only when registering new participants (#1268)
Helix's doesn't have an API for setting tags for instances atomically. They can only be added and removed one-by-one. Also, each tag modification could lead to rebalances of resources.

There are a couple of other issues with our current implementation:
1. Currently, we only add instance tags, but didn't purge ones that are no longer valid
2. We register a participant in Helix first, and then add the tags one by one. This will lead to multiple rounds of rebalances

This commit sets the tags in the instance config at the time new instances are being registered. We can edit the tags for existing instances operationally as a one-time effort.