[server][metrics] Fix sawtooth behavior on follower lag monitoring (#1305)
* [server][metrics] Fix sawtooth behavior on follower lag monitoring

For AA stores heartbeats from all peer colos are forwarded to the local follower replica, and the last received timestamp heartbeat is reported.  The effect of this is that if one colo is lagging, this will cause the metric to jump around as unhealthy heartbeats are interleaved with healthy ones.  This change addresses that by having the follower strictly report only the highest timestamp amongst heartbeats it's recieved.