[server] Increased Wait After Unsubscribe During State Transitions (#1213)
Increased the timeout value so that the consumer will more frequently safely unsubscribe instead of timing out. This should only be during partition state transitions because those are the times when not waiting for all inflight messages to be processed could result in a state mismatch. According to the values of the metric `consumer_records_producing_to_write_buffer_latency`, increasing the timeout to `1800s` should cover almost 100% of all cases.

Also added a log message for when `SharedKafkaConsumer#waitAfterUnsubscribe()` takes longer than 15 seconds to gather some data about how long it would take.