[server] Tuned zstd compression level in ingestion
Previously, Venice Server is always using the max compression level(22)
to compress the payload, which is very slow and this PR will change
the compression level to the default level (3) by default, and the
compression speedup is roughly 10-20x and it might introduce
some more storage overhead, 5-10%, but it should be fine.

This PR doesn't change the compression level in batch push as it is
offline, and if we see a need to speed up the map-reduce phase, we
can do the same later.

New Server Config:
server.zstd.dict.compression.level: default 3

New metric:
leader_compress_latency