[test] Make StoreIngestionTask::testStartServerAndShutdownWithPartitionAssignmentVerification & testCheckBeforeJoinCluster less flaky (#1330)
Make StoreIngestionTask::testStartServerAndShutdownWithPartitionAssignmentVerification less flaky by checking that the number of partitions is 0 instead of storageEngines is 0. This is because after the node comes online, Helix tries to re-assign partitions to it immediately, which re-creates the storageEngine.

Also made StoreIngestionTask::testCheckBeforeJoinCluster less flaky by turning on maintenance mode on the Helix cluster after removing the node from the cluster.