[server] swap map for list fields when examining idealstate for disk cleanup (#1335)
* [server] swap map for list fields when examining idealstate for disk cleanup
In a helix ideal state, helix maintains information in listfields and mapfields. The mapfields inform which messages needs to be sent based on the current LIVEINSTANCES, and will therefore be reflective of what nodes are available to receive messages. The listfields however will contain the assignment, and will only change if a rebalance is calculated. For this clean up code, we rely on entries in the listfield based on which state transitions we're anticipating to receive.