[controller] Allow configuring cloud information processor package in Helix CloudConfig (#1344)
Helix allows configuring "CloudConfigs" that can be used to populate instance "DOMAIN" fields on participant startup. In these configs, we can specify a class name and a package name. If a package name is not specified, Helix uses a default value of "org.apache.helix.cloud.<CLOUD_PROVIDER>". This might not be desirable for all usecases, so, this commit allows overriding the package name.