[router] Fix invalid tehuti sensor name passed in for latencies (#1351)
If an invalid TehutiMetricNameEnum that was not registered with MetricEntityState is passed in, it will be skipped during record() and will be silently ignored. Found this issue for healthy_request_latency and this PR fixes this issue and adds logging with redundant logging filter to log such issues. Not throwing an exception as we will be moving a lot of the existing metrics to Otel and tracking this log will help find and fix the issue during the migration and not affect the rollout process.