[test] Fix race condition in DaVinciClientTest (#1361)
There exists a race condition in DaVinciClientTest where tests with different schemas running concurrently can cause VeniceInconsistentSchemaException. This is because all of the tests in this file share the same directory, and the validation logic reads the first schema file in the directory.

The fix for this is to create a new temp directory for each test to avoid collisions.